[#30] Rename: post 폴더 이름 posts로 변경
