+++ src/components/Board/Board.tsx
@@ -6,7 +6,7 @@ import {
 import { headers } from 'next/headers';
 import { BoardCategory } from '@/types/board';
 import { fetchPosts } from '@/service/firebase/firebase-firestore';
-import Posts from '@/components/Board/post/Posts';
+import Posts from '@/components/Board/posts/Posts';
 import PopularPosts from '@/components/Board/popularPosts/PopularPosts';
 import WriteButton from '@/components/Board/write/WriteButton';
