+++ src/components/Board/posts/PostCard.tsx
