+++ src/components/Board/posts/PostCardContent.tsx
