+++ src/components/Board/posts/PostCardFooter.tsx
